requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
python-dotenv
requests
langchain
langchain_community
volcengine-python-sdk